requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
joblib
scikit-learn
pandas
numpy
tensorflow
pillow
passlib[bcrypt]
python-jose[cryptography]
